Move PluginCompat to give access to package.scala
