+++ src/main/scala/com/typesafe/sbt/web/package.scala
@@ -1,5 +1,7 @@
 package com.typesafe.sbt
 
+import PluginCompat.FileRef
+
 import java.io.File
 
 package object web {
@@ -12,5 +14,5 @@ package object web {
   /**
    * A function for possibly selecting a single file from a sequence.
    */
-  type Deduplicator = Seq[File] => Option[File]
+  type Deduplicator = Seq[FileRef] => Option[FileRef]
 }
